Remove color package, add fade and mix docs
