+++ __docs__/wonder-blocks-tokens/color-utils-fade.mdx
@@ -1,0 +1,69 @@
+import * as React from "react";
+import {Meta, Story, ColorItem} from "@storybook/blocks";
+import {View} from "@khanacademy/wonder-blocks-core";
+import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
+import {color, fade, font, spacing} from "@khanacademy/wonder-blocks-tokens";
+
+import equalsIcon from "@phosphor-icons/core/bold/equals-bold.svg";
+import plusIcon from "@phosphor-icons/core/bold/plus-bold.svg";
+
+<Meta title="Tokens / Color Utils / fade" />
+
+# fade()
+
+The `fade` function adjusts the alpha value of a color.
+
+## Usage
+
+```js
+import {fade} from "@khanacademy/wonder-blocks-tokens";
+
+fade(color: string, percentage: number): string;
+```
+
+## API
+
+| Argument     | TypeScript Type | Default    | Description          |
+| ------------ | --------------- | ---------- | -------------------- |
+| `color`      | `string`        | _Required_ | The base color       |
+| `percentage` | `number`        | _Required_ | The alpha percentage |
+
+## Example
+
+This example shows how to use the `fade` function to adjust the alpha value of a
+color. The `fade` function takes a color and a percentage as arguments and
+returns a new color with the adjusted alpha value.
+
+```typescript
+import {color, fade} from "@khanacademy/wonder-blocks-tokens";
+
+const fadedPurple64 = fade(color.purple, 0.64);
+const fadedGreen50 = fade(color.green, 0.5);
+```
+
+export const fadedPurple64 = fade(color.purple, 0.64);
+export const fadedGreen50 = fade(color.green, 0.5);
+
+<View
+    style={{
+        display: "grid",
+        gridTemplateColumns: "repeat(5, 1fr)",
+        gap: spacing.large_24,
+        justifyItems: "center",
+    }}
+
+>
+
+    <ColorItem colors={{color: color.purple}} />
+    <PhosphorIcon icon={plusIcon} size="large" aria-label="plus" />
+    <View style={{fontSize: font.size.xxxLarge}}>0.64</View>
+    <PhosphorIcon icon={equalsIcon} size="large" aria-label="equals" />
+    <ColorItem colors={{result: fadedPurple64}} />
+
+    <ColorItem colors={{color: color.green}} />
+    <PhosphorIcon icon={plusIcon} size="large" aria-label="plus" />
+    <View style={{fontSize: font.size.xxxLarge}}>0.5</View>
+    <PhosphorIcon icon={equalsIcon} size="large" aria-label="equals" />
+    <ColorItem colors={{result: fadedGreen50}} />
+
+</View>
