+++ __docs__/wonder-blocks-tokens/color-utils-mix.mdx
@@ -1,0 +1,70 @@
+import * as React from "react";
+import {Meta, Story, ColorItem} from "@storybook/blocks";
+import {View} from "@khanacademy/wonder-blocks-core";
+import {PhosphorIcon} from "@khanacademy/wonder-blocks-icon";
+import {color, mix, spacing} from "@khanacademy/wonder-blocks-tokens";
+
+import equalsIcon from "@phosphor-icons/core/bold/equals-bold.svg";
+import plusIcon from "@phosphor-icons/core/bold/plus-bold.svg";
+
+<Meta title="Tokens / Color Utils / mix" />
+
+# mix()
+
+The `mix` function allows you to mix a color into a background color, using the
+alpha channel of the base color to determine the linear blend.
+
+## Usage
+
+```js
+import {mix} from "@khanacademy/wonder-blocks-tokens";
+
+mix(color: string, background: string): string;
+```
+
+## API
+
+| Argument     | TypeScript Type | Default    | Description                                      |
+| ------------ | --------------- | ---------- | ------------------------------------------------ |
+| `color`      | `string`        | _Required_ | The base color                                   |
+| `background` | `string`        | _Required_ | The background color used to determine the blend |
+
+## Example
+
+This example shows how to mix a color with a background color. In this case, we
+are using an alpha channel of `0.32` for the base color, and a solid color for
+the background. The result is a color that is a linear blend of the two colors.
+
+```typescript
+import {color, mix} from "@khanacademy/wonder-blocks-tokens";
+
+const mixedColor = mix(color.offBlack32, color.white);
+const mixedColor2 = mix(color.white32, color.blue);
+```
+
+export const mixedColor = mix(color.offBlack32, color.white);
+export const mixedColor2 = mix(color.white32, color.blue);
+
+<View
+    style={{
+        display: "grid",
+        gridTemplateColumns: "repeat(5, 1fr)",
+        gap: spacing.large_24,
+        justifyItems: "center",
+    }}
+
+>
+
+    <ColorItem colors={{color: color.offBlack32}} />
+    <PhosphorIcon icon={plusIcon} size="large" aria-label="plus" />
+    <ColorItem colors={{background: color.white}} />
+    <PhosphorIcon icon={equalsIcon} size="large" aria-label="equals" />
+    <ColorItem colors={{result: mixedColor}} />
+
+    <ColorItem colors={{color: color.white32}} />
+    <PhosphorIcon icon={plusIcon} size="large" aria-label="plus" />
+    <ColorItem colors={{background: color.blue}} />
+    <PhosphorIcon icon={equalsIcon} size="large" aria-label="equals" />
+    <ColorItem colors={{result: mixedColor2}} />
+
+</View>
