+++ consistency-tests/tsconfig.json
@@ -9,7 +9,6 @@
         {"path": "../packages/wonder-blocks-button/tsconfig-build.json"},
         {"path": "../packages/wonder-blocks-cell/tsconfig-build.json"},
         {"path": "../packages/wonder-blocks-clickable/tsconfig-build.json"},
-        {"path": "../packages/wonder-blocks-color/tsconfig-build.json"},
         {"path": "../packages/wonder-blocks-core/tsconfig-build.json"},
         {"path": "../packages/wonder-blocks-dropdown/tsconfig-build.json"},
         {"path": "../packages/wonder-blocks-form/tsconfig-build.json"},
