--- packages/wonder-blocks-color/CHANGELOG.md
@@ -1,50 +1,0 @@
-# @khanacademy/wonder-blocks-color
-
-## 3.0.0
-
-### Major Changes
-
--   48d3c7e9: Fix how mix formats single-digit color components
-
-## 2.0.1
-
-### Patch Changes
-
--   d4c2b18c: Fix a variety of issues with Flow types generated by flowgen
-
-## 2.0.0
-
-### Major Changes
-
--   1ca4d7e3: Fix minor issue with generate Flow types (this is a major bump b/c I forgot to do one after doing the TS conversion)
-
-## 1.2.3
-
-### Patch Changes
-
--   b5ba5568: Ensure that flow lib defs use React.ElementConfig<> isntead of JSX.LibraryManagedAttributes<>
-
-## 1.2.2
-
-### Patch Changes
-
--   d816af08: Update build and test configs use TypeScript
--   3891f544: Update babel config to include plugins that Storybook needed
--   0d28bb1c: Configured TypeScript
--   3d05f764: Fix HOCs and other type errors
--   c2ec4902: Update eslint configuration, fix lint
--   2983c05b: Include 'types' field in package.json
--   77ff6a66: Generate Flow types from TypeScript types
--   ec8d4b7f: Fix miscellaneous TypeScript errors
-
-## 1.2.1
-
-### Patch Changes
-
--   91cb727c: Remove file extensions from imports
-
-## 1.2.0
-
-### Minor Changes
-
--   34c7aacb: Update the value of darkBlue from #0a2a66 to #0b2149
