--- packages/wonder-blocks-color/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@khanacademy/wonder-blocks-color",
-  "version": "3.0.0",
-  "design": "v2",
-  "publishConfig": {
-    "access": "public"
-  },
-  "description": "",
-  "main": "dist/index.js",
-  "module": "dist/es/index.js",
-  "types": "dist/index.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "devDependencies": {
-    "@khanacademy/wb-dev-build-settings": "^1.0.0"
-  },
-  "author": "",
-  "license": "MIT"
-}
