--- packages/wonder-blocks-color/tsconfig-build.json
@@ -1,9 +1,0 @@
-{
-    "exclude": ["dist"],
-    "extends": "../tsconfig-shared.json",
-    "compilerOptions": {
-        "outDir": "./dist",
-        "rootDir": "src",
-    },
-    "references": []
-}
