--- packages/wonder-blocks-color/types
@@ -1,1 +1,0 @@
-../../types
