+++ packages/wonder-blocks-dropdown/package.json
@@ -18,7 +18,6 @@
     "@babel/runtime": "^7.18.6",
     "@khanacademy/wonder-blocks-cell": "^3.3.2",
     "@khanacademy/wonder-blocks-clickable": "^4.1.2",
-    "@khanacademy/wonder-blocks-color": "^3.0.0",
     "@khanacademy/wonder-blocks-core": "^6.3.1",
     "@khanacademy/wonder-blocks-icon": "^4.0.1",
     "@khanacademy/wonder-blocks-layout": "^2.0.29",
