+++ packages/wonder-blocks-dropdown/src/components/option-item.tsx
@@ -2,8 +2,7 @@ import * as React from "react";
 import {StyleSheet} from "aphrodite";
 
 import {DetailCell} from "@khanacademy/wonder-blocks-cell";
-import {mix} from "@khanacademy/wonder-blocks-color";
-import {color, spacing} from "@khanacademy/wonder-blocks-tokens";
+import {mix, color, spacing} from "@khanacademy/wonder-blocks-tokens";
 import {LabelMedium, LabelSmall} from "@khanacademy/wonder-blocks-typography";
 
 import {AriaProps, StyleType, View} from "@khanacademy/wonder-blocks-core";
