+++ packages/wonder-blocks-form/tsconfig-build.json
@@ -8,7 +8,6 @@
     "references": [
         {"path": "../wonder-blocks-button/tsconfig-build.json"},
         {"path": "../wonder-blocks-clickable/tsconfig-build.json"},
-        {"path": "../wonder-blocks-color/tsconfig-build.json"},
         {"path": "../wonder-blocks-core/tsconfig-build.json"},
         {"path": "../wonder-blocks-i18n/tsconfig-build.json"},
         {"path": "../wonder-blocks-icon/tsconfig-build.json"},
