+++ packages/wonder-blocks-pill/src/components/pill.tsx
@@ -3,7 +3,7 @@ import {StyleSheet} from "aphrodite";
 import type {StyleDeclaration} from "aphrodite";
 
 import Clickable from "@khanacademy/wonder-blocks-clickable";
-import {mix} from "@khanacademy/wonder-blocks-color";
+import {mix} from "@khanacademy/wonder-blocks-tokens";
 import {View} from "@khanacademy/wonder-blocks-core";
 import {
     Body,
