+++ packages/wonder-blocks-pill/tsconfig-build.json
@@ -7,7 +7,6 @@
     },
     "references": [
         {"path": "../wonder-blocks-clickable/tsconfig-build.json"},
-        {"path": "../wonder-blocks-color/tsconfig-build.json"},
         {"path": "../wonder-blocks-core/tsconfig-build.json"},
         {"path": "../wonder-blocks-link/tsconfig-build.json"},
         {"path": "../wonder-blocks-tokens/tsconfig-build.json"},
