+++ packages/wonder-blocks-theming/tsconfig-build.json
@@ -6,7 +6,6 @@
         "rootDir": "src",
     },
     "references": [
-        {"path": "../wonder-blocks-color/tsconfig-build.json"},
         {"path": "../wonder-blocks-core/tsconfig-build.json"},
         {"path": "../wonder-blocks-tokens/tsconfig-build.json"},
     ]
