+++ packages/wonder-blocks-tokens/package.json
@@ -14,7 +14,6 @@
         "access": "public"
     },
     "dependencies": {
-        "@khanacademy/wonder-blocks-color": "^3.0.0",
         "@khanacademy/wonder-blocks-spacing": "^4.0.1"
     },
     "devDependencies": {
