+++ packages/wonder-blocks-tokens/src/index.ts
@@ -5,7 +5,19 @@ import {border} from "./tokens/border";
 import {color} from "./tokens/color";
 import {font} from "./tokens/font";
 
-/**
- * Core tokens for the Wonder Blocks design system.
- */
-export {border, color, font, spacing};
+import {mix, fade} from "./util/utils";
+
+export {
+    /**
+     * Core tokens for the Wonder Blocks design system.
+     */
+    border,
+    color,
+    font,
+    spacing,
+    /**
+     * Utility functions for working with colors.
+     */
+    mix,
+    fade,
+};
