+++ packages/wonder-blocks-tokens/tsconfig-build.json
@@ -9,9 +9,6 @@
     },
     "references": [
         {
-            "path": "../wonder-blocks-color/tsconfig-build.json"
-        },
-        {
             "path": "../wonder-blocks-spacing/tsconfig-build.json"
         }
     ]
