+++ tsconfig-build.json
@@ -12,7 +12,6 @@
         {"path": "./packages/wonder-blocks-button/tsconfig-build.json"},
         {"path": "./packages/wonder-blocks-cell/tsconfig-build.json"},
         {"path": "./packages/wonder-blocks-clickable/tsconfig-build.json"},
-        {"path": "./packages/wonder-blocks-color/tsconfig-build.json"},
         {"path": "./packages/wonder-blocks-core/tsconfig-build.json"},
         {"path": "./packages/wonder-blocks-data/tsconfig-build.json"},
         {"path": "./packages/wonder-blocks-dropdown/tsconfig-build.json"},
